Converting Settings to use the car-list lib
Now instead of holding TypedPagedListAdapter and its main LineItems
internally, Settings has a dependency on car-list and it pulls from a
more generalized version of that there.  The only main change to the
LineItems themselves is a universalized onClick(View view) that was
added to the base LineItem class called on the view being clicked.  One
new line item was added in SimpleIconTransitionLineItem which takes the
too specific behavior from SimpleIconLineItem and moves it out into
itself.

--Amend: Changed the AutoBrightnessLineItem test to use the new onClick
interface.  Still failing one of the tests that it was before but that's
out of the scope of this CL.

Bug: 65851280
Test: No changes in behavior, same tests as before
Change-Id: I84eb38307ec7689d2afe3db3078afd2512916a17