Making the not in range check consistent with phone
The way that we currently check to see if the AccessPoints returned by
the WifiTracker are in range is if their level is != -1.  This leads to
the issues with all of them occasionally saying Not in Range
temporarily.  This fix will make our check in line with the phone and
instead make them disappear temporarily.  This seems like a more
intuitive fix until I figure out the issue causing them all to
occasionally get the wrong config.

Test: make and manual test
Bug 67463612
Change-Id: I8e400e3c5d4e790b4f41fa64faefb580a04bcfdb