Car/Settings: Remove direct dep on sdk_vcurrent module.
Doesn't appear to be needed. If it was needed, the right way to depend on
this module is to specify LOCAL_SDK_VERSION. This particular module is
getting deleted soon.

Note that this test fails with and without my change.

Test: m -j RunCarSettingsRoboTests has no new failures
Bug: 77525052
Change-Id: I1a309cc29b04028ff38717ceb3503511a5a42d80
Merged-In: Id303080da52acd042201a490a5f7dade6a866d1e