Remove extras from BaseFragment.
This is a preliminary step in moving setting hierarchy into XML. It allows for most fragments to be instantiated with the default constructor and therefore launchable from the activity handling a clicked preference with a fragment attribute. Deferring removing "newInstance" methods that take parameters to subsequent refactoring steps.

All other changes are the result of running the formatter or bare minimum to pass upload hooks.

Bug: 112931076
Fixes: 78169047

Test: manually verify modified settings, make RunCarSettingsRoboTests, make RunCarSetupWizardRoboTests, adb shell am start -n com.google.android.car.setupwizard/.security.ChooseLockActivity
Change-Id: Ia9ed726d4218196115ed296aeb2bd4e713ba7162