DO NOT MERGE Updating the suggestion buttons to be generic
Removing the setup now button since the entire list item should open the
suggested action. Changed the dismiss button to have a generic dismiss
string and be capitalized in the strings file for translation purposes.
All of this can be updated after UX has time to finalize our decision
for this.

Change-Id: I02572e1e83d3a867398342323006925a557fd25e
Fixes: 77653293
Fixes: 111026792
Test: All passing