Add controller for injecting extra settings.
This migrates the existing behavior to use Preferences and does not seek to improve how extra settings are injected. A followup investigation is required to evaluate if the current strategy is robust, efficient, and meets the needs of clients. The duplication will be removed once all use cases are migrated (hence no attempt to share code between methods in ExtraSettingsLoader).

Bug: 112931076

Test: verified in end to end feature CL
Change-Id: Ie04d7a9471b9ff3fa843bff22e5897a668314ed2