Migrate ResetNetworkFragment to use preferences.
Note that phone settings does not use preferences for this setting. We do so here so that OEM styles are applied consistently across all settings screens. As part of the migration, this change includes missing features:
-The ability to reset eSIMs.
-The ability to select the network subscription.

The actual network reset occurs in the confirmation fragment which will handle reseting eSIMs and network subscriptions when migrated.

Bug: 117336671

Test: build and deploy, RunCarSettingsRoboTests
Change-Id: I9be31181780546792241a5303480ece7658c10b7