Migrate TimeZonePickerFragment to use Preferences.
This CL creates a TimeZonePreference type so that TimeZones can be
represented as preferences rather than list items.

There are two small bug fixes included in this change:
1. Previously the time zones were not sorted by the offset from GMT/UTC.
This change implements the TimeZonePreference as a Comparable so that
the Preference framework will automatically sort them. (Note that this
is the functionality we see in Phone settings).

2. This change also make a change that opens up the
TimeZonePickerFragment to put the current TimeZone at the top of the
screen. This functionality is also seen in phone settings but not the
current implementation of the car settings.

Note that we are not yet deleting TimeZoneListItem since master seems
to be having a hard time merging the delete. Will send a cleanup CL
afterwards.

Bug: 117335724
Test: Build, Manual Testing, Robolectric Unit tests for Preference and
Controller.
Change-Id: Ib86035dd9b29aafa5978abc45bc7991c3c6c49bc