Migrate DatetimeSettings to use Preferences.
UI has been changed to address b/115936025. Instead of using an extra
icon on the right of the preference, we are enabling/disabling the
preference instead.

This change also includes a fix where the description for 'Set time' and
'Use 24-hour format' gets updated on the ACTION_TIME_TICK broadcast
message (which triggers every minute). Previously, the description
remained at the time at which the Datetime fragment was opened.

While implementing, discovered a bug in the original UI where setting
'Automatic time zone' doesn't actually update the timezone. Will address
this bug in a separate change.

Bug: 117336350
Bug: 115936025
Test: Build, Manual Testing, Robolectric unit tests for controllers
Change-Id: Ia1ed895f98b8761a3be328e2ad8d3baf4bec731f