Separating the bluetooth dialog task from car settings
Car settings was automatically being brought to the front of the device
if it had been opened and then the user connects to bluetooth through
SUW. This seems to be caused because the dialog attaches to the
settings task. Added an affinity and the singleTask flag to the activity
so that this problem is fixed.

Fixes: 117828665
Test: Manual testing
Change-Id: I75db86353d7c375ce681adcc17e95584166eeb19