Migrate Display Settings to Preferences.
Display settings can be represented with SwitchPreference and
SeekBarPreference. Note that SeekBarPreference looks slightly different
than what originally existed (seek bar is not aligned with text), but
this can be changed via changing the preference styles in the future.

Also, actual testing of the change of the display brightness was not
observed on the Mojave, because there isn't a correct LightsService
built to handle it. Still, all of the logic remains the same as before.

We include our own version of SeekBarPreference due to the inflexibility
of the standard SeekBarPreference. We add a functionality to send
continuous updates to the listener while changing the seek bar.

Bug: 117336555
Test: Build, Manual, Robolectric unit tests
Change-Id: I4c1f96fcfc414ca77c1addec64ab76a4c5e434d2