Migrate Legal Information Settings to Preferences.
Note that there is a small change in UI. The supplemental icon on the
right hand side of the preference no longer exists, since Preferences
doesn't support such icons out of the box. This is consistent with what
exists on phone settings.

Third-party licenses also is broken (was already broken, due to "Webpage
at content://com.android.settings.files/my_cache/NOTICE.html could not
be loaded"). Will inspect in a separate CL.

Bug: 117336178
Test: Build, Manual, Robolectric unit tests
Change-Id: If81d3560fac47c1a7e2b4a71c59dc88b73477ca9