Prevent multiple calls to displayPreferences from duplicating items.
When we call screen.addPreference(), it adds the preference unless the
screen already contains the same Preference object. In
displayPreferences, we recreate the Preference object every time, which
results in duplicate entries.

Intead of keeping a list of ResolveInfo, we create the Preferences
within "populateApplicationList()". We keep this list of Preference objects
so that the screen doesn't think we have new items every time
displayPreference is called.

Bug: 117335481
Change-Id: I5053bb2bbc44486427b068957fc008e02d493a22
Test: Build, Manual, Additional unit test