Migrate ChooseLockTypeFragment to use Preferences.
As part of migrating everything to preferences, this change decouples
ChooseLockTypeFragment from the dialog which confirms the removal of the
current password.

This change also includes a change to CarSettingActivity, which allows
launchFragment() to also launch DialogFragments. Currently it passes an
empty tag. If the tag becomes necessary in the future (i.e. for restoring
after configuration changes) a simple solution is to create a
"BaseSettingsDialogFragment" which has a method to "getTag". Tag can be
overwritten in subclasses.

Bug: 117336419
Test: Manual, Build, Robolectric tests
Change-Id: I1fd8b4a3e90cf75adb85479b00145e9a20cea370