Migrate Language Settings to Preferences.
Language settings will be updated again to allow for multiple selected
languages as requested in b/117289809. This change focuses on migrating
the existing code to use Preferences.

Couple major changes:
1. ViewModel is no longer used. Language settings uses
LocaleStore.getLevelLocales() in multiple places to get list of
languages / list of child locales for given language / etc. The existing
code only used ViewModel to get the initial list of languages, but
wasn't consistent for the other calls to getLevelLocales. Since the
phone settings version does not use ViewModel and the requested list is
quite small, I decided to remove the usage of ViewModels in order to
have cleaner code.
2. Language settings is split into 2 separate fragments instead of just
using a single fragment and swapping out the data. This allows for less
intertwining of code between the Fragments and the Preference
Controllers and generally seems like better organization of code.

Bug: 117336750
Test: Manual, Build, Robolectric tests
Change-Id: I56e7526156e50cfefbfd5f49ce847d7fef08f1f3