Migrate UsersListFragment to use preferences.
The code mostly follows the existing structure. UsersPreferenceProvider
should fully replace UsersItemProvider, but it will be cleaned up in a
separate CL which migrates the ChooseNewAdminFragment.
UsersPreferenceProvider doesn't use a builder pattern anymore, since we
only need setters for 2 bools (rather than 4 properties before).

UsersListFragment still contains all of the logic relating to the Add
user button, which opens various dialogs based on certain properties
(i.e. demo or not). We might be able to clean up some of the dialog
logic, but that should be addressed in a separate CL. Some of the
listener interfaces implemented in the fragment were also moved into
the preference controller.

Note the we use a base class for the preference controller because most
of the logic will be shared with the choose new admin preference
controller in a subsequent CL.

Bug: 117336180
Test: Build, Manual, CarSettingsRobolectricTestRunner
Change-Id: I4b1fe1f0b9a8a7c8752692f286e8457332cce755