Migrate UserDetailFragment to Preferences.
This CL is quite large. It takes the UserDetailsFragment and splits it
into two separate fragments: UserDetailsFragment and
UserDetailsPermissionsFragment. These two functionalities were using
completely separate providers and such but only shared the App/Action
bar in common. As such, the common logic was extracted out into the
UserDetailsBaseFragment.

Both groups of preference controllers share a common logic in handling
the UserInfo and that logic was abstracted into
BaseUserDetailsPreferenceController.

Bug: 117336754
Bug: 119429161
Test: Build, Manual, Robolectric
Change-Id: Iaa81950497d8d4902373368e3dc9b660b5066ea1