Sound Settings connects to car on onCreate()
Sound settings used to connect to the car service on start
and disconnect on stop, but this was causing issues where
the number of sound preferences doubled every time we
natigated away from the sound settings screen without fully
destroying the screen.

Another way we could have addressed this bug is to just clear
the preference screen (and list of preferences) on onStop().
However, since the xml resources which provide the volume group
and sound usages are static, we thought it would be best not to
parse through these resources multiple times.

Bug: 119678736
Test: Manual, CarSettingsRobolectricTestRunner
Change-Id: Id4cbdda59166c4a3141ef6714c707b8addc6e6ba