Extend FragmentController to help with Dialogs
Previously dialogs were launched using
FragmentController#launchFragment. However this is restrictive because
it does not allow us to specify a tag to find the dialog in the future.

Dialogs are unable to restore state completely on configuration change
because the listener is not serializable/parcelable. In the chance that
a car switches day/night mode while a dialog opens, we should be able to
gracefully support it.

This change extends the fragment controller to include 2 additional
methods:
- showDialog
- findDialogByTag

showDialog acts exactly as launchFragment did to launch dialogs
previously. However it takes a 2nd nullable argument to specify the tag.

findDialogByTag purposefully restricts the fragmentmanager to finding
fragments of type DialogFragment. This is to discourage users from using
this functionality in random cases.

Bug: 119820132
Test: Robolectric tests, Manual (open lock screen dialog, trigger
configuration change, test the listener is triggered)
Change-Id: Ic6a82940ef9fe0a47439ed6c4b1806c9b86aef54