Add preferred engine setting for TTS
This CL adds the first setting for TTS, which is to set the preferred
engine. It differs slightly from the phone settings. Phone settings uses
radio buttons to signify that a TTS engine is the current engine. This
implementation simply uses the preference summary to do so.

One feature that needs to be added for this setting is a dialog which
alerts the user that the TTS may collect their data and requires the
user to acknowledge it. In order to keep CLs smaller, I will add this in
a separate subsequent CL.

Writing tests was fairly difficult due to the way TextToSpeech.java was
written. It relies heavily on the implementation of ShadowTextToSpeech.
Any suggestions for improvment is welcome.

Bug: 117290240
Test: Build, Manual (install Vocalizer TTS from play store and try
switching default), Robolectric test
Change-Id: I741f46053317683b84bc563bf49ab418cbf12c57