Fix issue for granting admin access.
The bug was introduced when we stopped using launchFragment() to show
dialogs in Settings app code. This was so we can properly use TAGs so
that dialog listeners could be reset after configuration change.

This change fixes the specific dialog which should confirm granting
admin access. It also fixes this dialog so that the listener is reset on
configuration change. A snippet of code has been added to launchFragment
so that this issue doesn't come up again (or is easier to find).

Bug: 120301742
Test: Manual, Build, Robolectric
Change-Id: I16e6c6427fe9649b7917e71c84ac4e8917948c49