Create Developer options settings page.
This change creates a Preference to open the Developer options screen.
This preference is only visible if developer options has been enabled.

The developer options screen is empty, except for a toggle in the action
bar, which allows you to disable developer options if so desired. When
enabling developer options from the action bar, it throws up the same
dialog as the phone settings version.

Bug: 117289448
Test: Build, Manual, Robolectric
Change-Id: I51a96ac515099e3062798fb2a2029c9e4e5cd071