Migrate add wifi fragment to preferences
Note that the entire fragment looks different. However the main goal of
this CL is to remove the dependencies to ListItemSettingsFragment, which
is a prereq to properly theming the Settings app. Additional components
could be built in order to bring it to look and feel the same as before,
but since Wifi settings features has changed significantly in Q, it is
worth seeing what is going to happen there before investing significant
time here.

Bug: 117337407
Test: Manual, Build, Robolectric
Change-Id: I5c5d9d3e3e44efaa46f154137b9acbbd27ce5b1b