DO NOT MERGE Add dialogs to enable renaming the local and remote Bluetooth devices.
This is one of many individual changes to migrate Bluetooth settings to the Preference framework. For end to end usage, see ag/5258452.

The alternative of using dialog preferences for this purpose was investigated, but difficulty in customizing the dialog edit text to restrict name length stalled that option. It may be worth investigating again in the future for code consistency purposes.

Bug: 117334982
Bug: 117335483

Test: RunCarSettingsRoboTests, tested end to end in full feature CL.
Change-Id: I4a917d45c739f983eace4431ff3ceaca2dd1b9dc