Update BluetoothSettingsFragment to use the new preference UI.
This change will enable the new Bluetooth UI submitted in recent CLs. Existing functionality such as swipe to refresh and the progress bar are now covered by other components. The switch is also disabled while the adapter is turning on or off and for restricted users to match the phone Settings behavior. A following CL will clean up unused code from the old implementation.

Bug: 117334982

Test: build, deploy, pair/unpair, connect/disconnect etc. RunCarSettingsRoboTests
Change-Id: Id55269a6df5271f7b75785cc0c7937df46e04e01