Add MasterClearFragment and MasterClearConfirmFragment.
This makes use of recently submitted preference controllers to complete the factory reset user flow. On the initial fragment, a user is presented with information about what will be reset. They must then authenticate and confirm again on a final warning page. The majority of the reset is handled by another system component, but under certain circumstances, the settings app is responsible for clearing persistent data. Currently, clearing external storage is considered out of scope for the automotive case.

Bug: 77476980

Test: build and deploy, reset device, RunCarSettingsRoboTests
Change-Id: Ibaae40917f0bde2c3ae20d80b66a79ccf5ba029b