Back button invisible on SUW lock screens
It appears that some changes to the coloring resulted in the back button
being the same color as the background. Since the icon base color is
black there is no way to tint it lighter. So for that reason I created a
separate icon to be used for setup wizard only.

Fixes: 122675811
Test: Ran setup wizard and could see the back button in lock screen

Change-Id: I748aa2e93ced963526c830a0c3f394605e3e6462