Add MasterClearAccountsPreferenceController.
This controller will list the currently signed in accounts (if present) for all profiles of the current user to inform that they will be removed during a factory reset. This is one of several CLs which when submitted will enable factory reset (aka "master clear") in CarSettings. Usage can be seen in ag/6258577

Bug: 77476980

Test: RunCarSettingsRoboTests, end to end in full feature CL.
Change-Id: Id54ea0391633b59cc337820cb1b60bb09bc9a0e3