DO NOT MERGE Implement uninstall apps in CarSettings.
This follows as close as possible to the constraints implemented in Phone Settings. We don't yet have a mechanism for OEMs to define "keep enabled" packages, but it can be included in a later change. Phone Settings redirects to Device Admin pages which are not implemented in CarSettings for certain uninstall conditions. We disable the button for now pending adding those pages / direction from UX.

Bug: 79492615

Test: build and deploy, install/uninstall/enable/disable/force stop, RunCarSettingsRoboTests
Change-Id: I3a3f0a52eeb34f0b8708210a25b47f3ea75935e1