Retain value of mHasNewIntent across configuration changes.
Right now, it's resetting to the default (true) which means that the current fragment is replaced with the FragmentResolver default (quick settings) when we receive a config change. This behavior didn't occur before because the FragmentResolver used to return null by default which would be rejected by launchIfDifferent(...) (updated in ag/6226655).

Bug: 121285423
Bug: 122726972

Test: manually send intents, toggle day/night, RunCarSettingsRoboTests
Change-Id: I13f300ad0ed84d06a7bd13a9a71c47252ea91fb8