Keep scanning while on the Bluetooth pairing page.
Bluetooth discovery times out after a set interval, but we want to keep searching for devices until the user has selected one to pair.  This is consistent with the Phone Settings behavior.  This change also sets the progress bar to show as long as the page is open (vs. only when scanning).  When scanning timed out, there was a slight delay before it could be restarted resulting in a flicker and poor user experience.  Since the bonding process is still making progress for the user's task, showing the bar seems reasonable even when not scanning.

Fixes: 122473126

Test: RunCarSettingsRoboTests, build, deploy, complete several pairing cases.
Change-Id: Id319b45e7d69b405c494bed2c425316e47cef7e9