Change default voice input code to conform to phone settings behavior
There are some parts of the phone settings code that was questionable
(i.e. commented out code, booleans that are always true) which I had
tried to "fix" previously. However, on second thought, it should be
better to conform to the behavior of phone settings.

In order to do so, we're updating the way that the voice input service
is shown as a pickable option or a disabled option.

We're also allowing both the recognizer and related interactor to be
shown by the VoiceInputInfoProvider.

Bug: 111550192
Test: Robolectirc, Manual
Change-Id: I3c1c6175c23e82e1bba8e0bfe818a1e4e36c78c4