See overall mobile data usage
This change makes quite a bit of changes to the phone settings
implementation. It currently only supports mobile data, while the phone
settings implementation also can support wifi usage or ethernet usage.

Also the current implementation tries it's best to separate the UI from
the business logic by using a plain ProgressBarPreference and doing all
of the logic in the preference controller. As such the logic in the
current preference controller comes from both DataUsageSummaryPreference
and DataUsageSummaryPreferenceController in the phone settings world.

Bug: 117522484
Test: Manual (in Jeep), Robolectric
Change-Id: I0700097064d47f896a72b64ea9e6d035a997b53e