Mock UX restrictions earlier in setup for CarSettingActivityTest.
ag/6662218 uncovered this error because it changed ACTION_SETTINGS to show HomepageFragment which would query UX restrictions (causing a null pointer exception). The UX restriction must be set before the activity is created.

Also switching to a different fragment for testing as the HomepageFragment requires additional dependency mocking not relevant to the activity/test.

Bug: 127950533

Test: RunCarSettingsRoboTests
Change-Id: I38eadadee1529f9844b6f7ab9d77d015ca25f766