Add additional intents that we need to handle.
This covers most of the rest of what we plan to support for V1 (for
intents which redirect to a particular page).

The remaining few have to do with Special App access settings and About
settings.

We have some additional intents that open dialogs which end up
performing actions. This will need a little bit more thought.

Bug: 114008723
Test: manual
Change-Id: Id3aa4e82545bb42bc9390071b52ffca1f09b2776