DO NOT MERGE Empty change to assist in merging keyguard-refactor-pi-dev topic.
ag/6730824 cannot automerge due to branch divergence, but the pathway between branches ignores "DO NOT MERGE". ag/6841626 is the manual cherry-pick, but the change is dependent on the rest of the topic merging. This CL will provide an ID to mark ag/6730824 as merged so that automerger doesn't make an attempt.

Bug: 120484642

Test: N/A
Change-Id: I20b050dd80d189516acb0baa63bc2483c7d36667