Rename car:text attribute to prevent overriding appompat
The app compat library defines a resource "attr/text" with format="string"
while CarSettings defines a resource of the same name with format="reference".
The appcompat resource is being overriden by the CarSettings resource and is
preventing aapt2 from disallowing multiple definitions of the same attribute
with different formats.

This change renames the CarSettings text attribute so it will not
collide with the appcompat resource.

Bug: 129146717
Test: m -j CarSettings && atest RunCarSettingsRoboTests
Change-Id: Ic442936e7f9e13dd149e6fbff3981b882e7ce689