DO NOT MERGE Adding setup lock setting service
Adds support for setting the lock through setup wizard if there is not
one already set. This includes making it so that this service is
provided ONLY to setup wizard and will give it all the functionality to
match its existing implementation in settings. The removal of the old
settings implementation and dependency on the primary Car support
library will be next.

Fixes: 129551356
Fixes: 131248006
Test: Robolectric tests and manual testing, all passing.

Change-Id: Ic2421e47f99e1dce50c32939dcf72025a6d16dba