Create ValidatedEditTextPreferenceDialogFragment to add validation to the provided input.
Also separated PasswordETPDF into a separate class in order to simplify ETPDF.

ETPDF allows pressing Enter to submit the Dialog. It should be used when no additional functionalities described below are needed.
VETPDF extends ETPDF and should be used in conjunction with ValidatedEditTextPreference when input can be submitted only if it is valid. Examples include: Setting the Hotspot name that cannot be empty or too long.
PETPDF extends VETPDF and should used in conjunction with PasswordEditTextPreference when the input should be obscured by default. It adds to the Dialog a checkbox that toggles the input visibility. Examples include: Entering the password to connect to WiFi.

Bug: 113169960
Test: Manual and Robolectric
Change-Id: Ic3c3e2dafaf2069fc7690c0e9f3edb4bdae28bc7