SettingsFragmentTest fix
About a week ago, the root fragment name check against the config value was added for determining whether to show the back icon or the application icon in the action bar. Along with that change, we started tagging fragments to retrieve the root fragment by its tag for the said comparison.

This change did not make it into BaseTestActivity used for tests, which led to the action bar always returning back icon since the root fragment couldn't be retrieved with a tag.

bug: 132073547
Test: run robolectric test.
Change-Id: Id71a90ff2bb73dee0cce7aac623a0e2acd3e8df5