Allow OEMs to show/hide action bar icon of setting's root through config
This required a new boolean configuration item whose value would be read
during onActivityCreated of SettingsFragment and the action bar icon
would only be hidden if the fragment name matched the settings root
fragment config item. Similar logic was also applied in
QuickSettingsFragment as that is the default settings root but doesn't
extend SettingsFragment.

Bug: 131853723
Test: Manual
Change-Id: I944dc9c0468e9878fcbff21d455890bed7da4b07