Scroll to a newly connected access point.
If the list of access points is long and the user selects to connect to
one that's several pages down, the access point will jump to the top of
the list, but the list itself will stay centered on the previous
location. Scrolling to the preference makes it so the user can see the
connection was successful.

Fixes: 131233898

Test: connect to access point, ensure list goes to top.
Change-Id: Ia4da1dde4e8ca8ac89255613ea5cf0eb04bc92e2