Move progress bar so it doesn't move page content.
Because the progress bar is switched from gone to visible, it shifts
list content when it is shown. Moving the progress bar to the same
container prevents this. There are other options but they have
drawbacks:
-Switch to invisible: leaves extra space when it is not needed
-Make part of the action bar: hard to keep in sync if added to
individual layouts, needs to be set in container each time action bar is
inflated if not.

Fixes: 135639679

Test: build and deploy, inspect Wi-fi and Bluetooth pages while loading.
Change-Id: Ib802c451e1ba8405b875d1f1ff3132210c27c962