Small fixes to add account flow
When creating a new instance of the AuthenticatorHelper, ensure that we
unregister and reregister the listeners to point to the same listener.
Otherwise we would leak an intent filter.

Also, there are some issues when processing activity results which
involve navigation, if we perform such action while Settings app is
stopped

Bug: 137897567
Test: manual, existing robolectric tests
Change-Id: I3d0e4d366834fb6968a82dbc1693e88ef709a2f6