Finish activity after terminateEnrollmentHandshake
When there is overlay view pop up during the enrollment, we will
terminate the enrollment but the loading page(or the verification code
dialog) will still remain on the background which will confuse the user
that the enrollment process can assume, so thisnCL will directly finish
the activity after call terminateEnrollmentHandshake. So what the user
will see is that when certain overlay view pop up, the background will
be back to "Add trusted device" button page.

Bug:138339107
Test: Build and flash
Change-Id: I39d20c0bea02b743860d3be144c31e6346922f86