Rework the intent handling logic
Recent changes removed some logic regarding checking the fragment
against the current fragment, which caused some regressions around
performance.

The current logic doesn't handle one specific case that was covered
before. If a user had opened Settings previously before to the Account
Settings before, navigated away and pressed a button which intented into
the same page, we do not clear the backstack as we might expect. However
this seems like enough of an edge case to move forward with.

Bug: 141203096
Test: manual, Robolectric
Change-Id: Id2ff95275049fc5067a880c401d10d0bf3700d61