Refactor OnUsersUpdateListener
OnUsersUpdateListener was a convenience listener to register for all
possible user events.  However, the API was not that customizable and so
it tends to fire for too many events.  Also, because it grouped so many
intents together, listeners lost all possible metadata associated with
the intents.

Rather than maintain this large list, we will instead remove the API and
have the callers register for these signals themselves.

Bug: 137136907
Test: atest CarSettingsRoboTests
Test: atest CarUserManagerHelperTest

Change-Id: I958a100aa832c809c44834dd009dc5509166263a