Update user event listeners
Follow up to ag/9484624. This CL cleans up the broadcasts that each of
our controllers is using to listen for user events. Since Settings is a
per-user app, it's not required to listen for ALL user events, only the
events of the currently running user. Similarly, reduced the intents
registered to just the ones needed.

Also fixed a bug introduced in ag/9484624 where we were crashing due to
unregistering a BroadcastReceiver that wasn't registered.  Turns out
some subclasses were not calling super.onCreateInternal();

Bug: 142019015
Test: atest CarSettingsRoboTest
Test: Manually add/remove users and switch to users/guest from quick
settings user picker
Test: Manually add/remove users from Settings user page

Change-Id: Ib404cc803ce0ba9a206991a714f2d45c2eac5c2c