Changed FallbackHome to finish itself more aggressively.
FallBackHome is the initial Home activity until the user is unlocked, at which
point it "kills itself" so the framework can launch the proper, user-specific
Home app.

Usually the activity dies soon after calling finish(), but often it takes a
while (specially when the device has activities running in a secondary display),
which can slow down the perceivable user switch on cars.

This CL mitigates the delay by calling finishAndRemoveTask() instead of finish();
this method is more efficient, and it's safe to use in this case (as the stack
should go away anyways)

Bug: 140999054
Test: manual verification by creating a new user, manually switching back and
      forth between the 2 users and observing that the home app is rendered
      quickly after the user is unlocked.

Change-Id: I79b893a039927b88c0a293285abcba50f898918e
(cherry picked from commit 1138d787930f49d1c114d20361b2258b18bee25b)