Changed ExtraSettingsPreferenceController's way of reading metadata boolean. Fixed flaky unit test.
Previous way of reading the xml metadata is incorrect. If the key does
not exist, the metadata will return a string. If the key exists, the
metadata will return a boolean. So this code needs to be updated so it
accounts for both.

Bug: 148946454
Test: Robolectric, manual
Change-Id: Ic16667632bc146f53a42cad0bbec8a130c82c7f1
Merged-In: Ic16667632bc146f53a42cad0bbec8a130c82c7f1