Rotary direct manipulation of SeekBar prefs
Support direct manipulation (DM) of SeekBar preferences using the
rotary controller. This is done using the advanced direct manipulation
mode. Basic direct manipulation doesn't work because the focus is on
the entire preference, not the SeekBar within it. To indicate DM mode,
the entire preference and the SeekBar within it are selected. This
causes the preference to appear unfocused and the SeekBar's thumb to
appear focused.

Bug: 161484085
Test: manual
Change-Id: I54bdde9203f10718354d6844cee5225107468b92