Revert^2 "Don't launch the real home for User 0 in Headless mode."
This reverts commit 3c9fbb5d8b841e817bc5029a6f91b26f1336e947.

Reason for revert: Roll forward the orginial CL with the fix of the package name change.

=== Original CL ===
Don't launch the real home for User 0 in Headless mode.

During the boot process, the launcher for User 0 can be shown unexpectedly, we won't launch User 0's read home to avoid it.
After boot, when switching to User 0, FallbackHome will not be called, since User 0 is already unlocked.

Bug: 166670668
Test: Factory reset and check if the launcher is shown before SUW.
Change-Id: Ia6bb117d12e5967410d99508c5fc0620f486b455