Fix Back button behavior
SeekBarPreference's event handler was consuming Back button presses when
in direct manipulation mode as well as when not in direct manipulation
mode. This prevented the Back button from navigating back when a
SeekBarPreference was focused. Now the event handler only consumes the
events when in direct manipulation mode, allowing the Back button to
navigate back when not in direct manipulation mode.

Test: press Back with SeekBarPref focused, press Back in DM mode
Bug: 170443437
Change-Id: Ic8136270cf693564838140320ad8e21ea143010b