Fix certain preferences not displaying properly
This change is to fix an issue where wifi entry and bluetooth devices
aren't showing properly. The issue started happening when we started
using ConstraintLayout, and only happens in single-pane mode where the
top_level_menu and divider view visibilities are set to gone. I'm
replacing it with a RelativeLayout as a workaround.

Bug: 182810085
Test: Observed the wifi and bluetooth screens in single and dual-pane
modes, as well as in rtl.

Change-Id: I4bf6bd3358b48fb6058126abd186e17a3d1743ad