Remember bluetooth profile when reconnecting
Currently to get around the fact that bluetooth devices will never
connect when phone and media profiles are off, both profiles are forced
on when connecting. Now the logic is being changed to only force them
on when both are disabled, so combinations such as phone off, media on
are remembered.

Bug: 190549594
Test: Manually connected/disconnected with various profile combinations
and made sure they are the same before and after reconnecting except for
when both phone and media are toggled off (then they are both forced
on). Also ran atest CarSettingsUnitTests

Change-Id: I5ebbd078a93e1d27de5c9667ce4218bb8f9599f6