Load number of unused apps in the background
The phone team flagged that this operation can take hundreds of
milliseconds. This change shows the item immediatley and loads the count
for the summary in the background.

This also includes the following changes:
- Always show the "Unused apps" setting item. The phone team also decided
to always show this item, instead of only showing when the count is
greater than 0. This is because otherwise the UI might skip around with
the background loading.
- Fix a bug with how the number of hibernated apps is counted. Copied
changes from ag/14471254

Bug: 191922683
Test: atest HibernatedAppsPreferenceControllerTest
Test: atest HibernatedAppsItemManagerTest
Change-Id: Ic80750eeb0b8dee639a78f9f868e79e437807643