Adding closed captions style preset settings to accessibility settings
This uses the settings secure constant to choose one of the 5
preconfigured choices for captions settings within Android. This is the
same as the functionality on the phone and the settings will be picked
up by video apps which use platform caption functionality.

Also added captions intent handling to the captions page and updated
Accessibility manifest entry to the correct activity.

Bug: 202214182
Bug: 202212154
Test: Added robolectric tests and tested locally
Change-Id: I8f5826ffd8afb94a0ec04ef8de52a32d3e9bdbf4