Add screen reader settings to CarSettings
This creates the screen reader section which allows a single
specific screen reader which can be overridden with an overlay.
The section will only be displayed if the screen reader is
installed on the device. The only setting for screen readers
is to enable/disable it as a switch.

Bug: 203584599
Bug: 205159618
Test: Manual/robolectric
Change-Id: I40847d05cc093f2b659a9191491c947720311d97
Merged-In: I40847d05cc093f2b659a9191491c947720311d97