Fixed Volume Mute Setting Test
Mute settings test was failing due to a check for the car audio service
mute feature. Added the correct check return to the audio manager mock,
this way the test can correctly run. Also added a test for the case
when the feature is disabled, in this case the volume mute should not
be updated.

Bug: 228112063
Test: atest VolumeSettingsPreferenceControllerTest
Change-Id: I3f87535d7c6ea8e00f133e223ef73ccdafbeccde