Change manual date/time limit handling
This commit changes how limits are applied to the manual date / time
settings UI.

Before this change there were no explicit limits set, so the user could
select all dates from 1900 - 2100 (which are the defaults for every date
picker as far as I can tell). However, if the date chosen was before a
hard-coded time in Nov 2007, the hard-coded lower bound would be set by
the settings app. The user would not be told this "flooring" behavior
had happened.

On mobile devices, this behavior can cause confusion, especially for
testers and it is assumed the same is true for Automotive.

With this change, the bounds are explicitly set from constants obtained
from an internal API (TimeDetectorHelper). At the time of writing the
year bounds are 2008 - 2100 on Android Automotive.

[Since all automotive devices are 64-bit only, the upper limit returned
on 32-bit or 32/64-bit hybrid devices of 2037 to avoid the Y2038 issue
is not used.]

This change means that the user is not able to set a date/time that the
system will not accept. Further, the bounds can easily be changed in
future by changing logic / constants in TimeDetectorHelper.

Bug: 239981321
Test: Manual on seahawk and atest CarSettingsUnitTests:com.android.car.settings.datetime.DatePickerFragmentTest
Change-Id: I98dc1130e12fa9181c01647b05c507fcdf86ad1e