Changes button layout for the bluetooth device preference.
In order to improve UI, shows single connect bluetooth button when device is not connected. If the device is already connected or going to be connected as a result of clicking connect button, shows phone and media profile buttons.

Change-Id: I16dec2be09352cb23c5a4b147c220c72a2516490
Bug: 246286977
Test: tested on seahawk atest BluetoothBondedDevicesPreferenceControllerTest
Test: tested on seahawk atest BluetoothDevicePreferenceTest
Test: tested on seahawk atest MultiActionPreferenceTest
Merged-In: I16dec2be09352cb23c5a4b147c220c72a2516490